Move initMjReceiver into BC proxy
