+++ common/logisticspipes/proxy/interfaces/IBCProxy.java
@@ -1,9 +1,12 @@
 package logisticspipes.proxy.interfaces;
 
 import logisticspipes.asm.IgnoreDisabledProxy;
+import logisticspipes.blocks.powertile.LogisticsPowerJunctionTileEntity;
 import logisticspipes.pipes.basic.LogisticsTileGenericPipe;
 import logisticspipes.proxy.buildcraft.subproxies.IBCPipeCapabilityProvider;
 import logisticspipes.recipes.CraftingParts;
+
+import javax.annotation.Nonnull;
 
 public interface IBCProxy {
 
@@ -25,4 +28,6 @@ public interface IBCProxy {
 	void registerInventoryHandler();
 
 	IBCPipeCapabilityProvider getIBCPipeCapabilityProvider(LogisticsTileGenericPipe pipe);
+
+	Object createMjReceiver(@Nonnull LogisticsPowerJunctionTileEntity te);
 }
